+++ src/onegov.people/onegov/people/upgrade.py
@@ -3,6 +3,7 @@ upgraded on the server. See :class:`onegov.core.upgrade.upgrade_task`.
 
 """
 from onegov.core.upgrade import upgrade_task
+from sqlalchemy import Column, Text
 
 
 @upgrade_task('Rename academic_title to salutation')
@@ -10,3 +11,9 @@ def rename_academic_title_to_salutation(context):
 
     context.operations.alter_column(
         'people', 'academic_title', new_column_name='salutation')
+
+
+@upgrade_task('Add function column')
+def add_function_column(context):
+    context.operations.add_column(
+        'people', Column('function', Text, nullable=True))